18662: Fixed - 'Disassembler' has not been declared
Backport-of: fdd58214f8e9547ab87f32a3387931ae9d831e34
