+++ src/hotspot/cpu/loongarch/assembler_loongarch.cpp
@@ -409,7 +409,6 @@ void Assembler::ld_w(Register rd, Address src){
       ldx_w(dst, base, AT);
     }
   }
-//Disassembler::decode(pc()-32, pc(), tty);
 }
 
 void Assembler::ld_wu(Register rd, Address src){
