+++ src/hotspot/cpu/mips/assembler_mips.hpp
@@ -967,11 +967,11 @@ public:
     AbstractAssembler::flush();
   }
 
-  inline void emit_long(int);  // shadows AbstractAssembler::emit_long
-  inline void emit_data(int x) { emit_long(x); }
-  inline void emit_data(int, RelocationHolder const&);
-  inline void emit_data(int, relocInfo::relocType rtype);
-  inline void check_delay();
+  void emit_long(int);  // shadows AbstractAssembler::emit_long
+  void emit_data(int);
+  void emit_data(int, RelocationHolder const&);
+  void emit_data(int, relocInfo::relocType rtype);
+  void check_delay();
 
 
   // Generic instructions
@@ -1785,8 +1785,5 @@ public:
   static address locate_operand(address inst, WhichOperand which);
   static address locate_next_instruction(address inst);
 };
-
-
-#include "assembler_mips.inline.hpp"
 
 #endif // CPU_MIPS_VM_ASSEMBLER_MIPS_HPP
